requirements.txt
boto3==1.20.8
certifi==2021.5.30
future==0.18.2
imageio==2.9.0
joblib==1.0.1
numpy==1.21.4
opencv-python==4.5.3.56
Pillow==8.3.2
scikit-learn==0.24.2
scipy==1.7.2
threadpoolctl==2.2.0
wheel==0.37.0
torch==1.6.0
torchvision==0.7.0
